Move clientBase and nodeClient to common, because the NodeClient is also used in the CLI
