+++ vantage6-client/vantage6/client/algorithm_client.py
@@ -4,7 +4,7 @@ import time
 
 from typing import Any
 
-from vantage6.client import ClientBase
+from vantage6.common.client.client_base import ClientBase
 from vantage6.common import base64s_to_bytes, bytes_to_base64s
 from vantage6.common.task_status import TaskStatus
 from vantage6.tools.serialization import serialize
